Move Constraints under Routing namespace
